refactor(imports): update module paths after files migration
