+++ app/core/auth.py
@@ -1,10 +1,10 @@
 from flask import request, g
 import jwt
 from functools import wraps
-from model.models import User
-from model.sessions import get_session
-from config import Config
-from core.errors_handlers import UnauthorizedError, ForbiddenError
+from app.model.models import User
+from app.model.sessions import get_session
+from app.config import Config
+from app.core.errors_handlers import UnauthorizedError, ForbiddenError
 from typing import Callable, Any
