+++ app/model/database.py
@@ -1,4 +1,4 @@
-from config import Config
+from app.config import Config
 from sqlalchemy import create_engine
 from sqlalchemy.engine import Engine
 from sqlalchemy.orm import sessionmaker, declarative_base, Session
