+++ app/model/sessions.py
@@ -1,5 +1,5 @@
 from flask import g, Flask
-from model.database import SessionLocal
+from app.model.database import SessionLocal
 from sqlalchemy.orm import Session
 from typing import Optional
